moved drawer menu and top bar menu to menu folder for fast accessbility, and changed their paths where were used
